+++ CHANGELOG.md
@@ -11,11 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `redirect_stdin` class in common.py (#10)
+- `remove_docstrings` function in audit.py
 
 ### Fixed
 
 - shorten output for complex errors
 - separate install for venv and jq
+
+### Removed
+
+- remove_comments.py -- now in audit.py
 
 
 ## [1.3.0] - 2025-09-17